refactor(router): move route handlers into their own folder
