+++ modules/angular2/src/router/route_handlers/async_route_handler.ts
@@ -1,7 +1,7 @@
 import {isPresent, Type} from 'angular2/src/facade/lang';
 
 import {RouteHandler} from './route_handler';
-import {RouteData, BLANK_ROUTE_DATA} from './instruction';
+import {RouteData, BLANK_ROUTE_DATA} from '../instruction';
 
 
 export class AsyncRouteHandler implements RouteHandler {
