+++ modules/angular2/src/router/route_handlers/route_handler.ts
@@ -1,5 +1,5 @@
 import {Type} from 'angular2/src/facade/lang';
-import {RouteData} from './instruction';
+import {RouteData} from '../instruction';
 
 export interface RouteHandler {
   componentType: Type;
