+++ modules/angular2/src/router/rule_set.ts
@@ -11,8 +11,8 @@ import {
   PathMatch
 } from './rules';
 import {Route, AsyncRoute, AuxRoute, Redirect, RouteDefinition} from './route_config_impl';
-import {AsyncRouteHandler} from './async_route_handler';
-import {SyncRouteHandler} from './sync_route_handler';
+import {AsyncRouteHandler} from './route_handlers/async_route_handler';
+import {SyncRouteHandler} from './route_handlers/sync_route_handler';
 import {Url} from './url_parser';
 import {ComponentInstruction} from './instruction';
